+++ DelvUI/Interface/GeneralElements/LabelHud.cs
@@ -1,5 +1,6 @@
 ﻿using System.Numerics;
 using Dalamud.Game.ClientState.Objects.Types;
+using Dalamud.Interface;
 using DelvUI.Config;
 using DelvUI.Helpers;
 using ImGuiNET;
@@ -24,7 +25,7 @@ namespace DelvUI.Interface.GeneralElements
             Draw(origin);
         }
 
-        public void Draw(Vector2 origin, Vector2? parentSize = null,
+        public virtual void Draw(Vector2 origin, Vector2? parentSize = null,
             GameObject? actor = null, string? actorName = null, uint? actorCurrentHp = null, uint? actorMaxHp = null)
         {
             if (!Config.Enabled || Config.GetText() == null)
@@ -39,7 +40,7 @@ namespace DelvUI.Interface.GeneralElements
             DrawLabel(text, origin, parentSize ?? Vector2.Zero, actor);
         }
 
-        private void DrawLabel(string text, Vector2 parentPos, Vector2 parentSize, GameObject? actor = null)
+        protected virtual void DrawLabel(string text, Vector2 parentPos, Vector2 parentSize, GameObject? actor = null)
         {
             using (FontsManager.Instance.PushFont(Config.FontID))
             {
@@ -92,6 +93,58 @@ namespace DelvUI.Interface.GeneralElements
                 default:
                     return Config.Color;
             }
+        }
+    }
+
+    public class IconLabelHud : LabelHud
+    {
+        private IconLabelConfig Config => (IconLabelConfig)_config;
+
+        public IconLabelHud(IconLabelConfig config) : base(config)
+        {
+        }
+
+        public override void Draw(Vector2 origin, Vector2? parentSize = null,
+            GameObject? actor = null, string? actorName = null, uint? actorCurrentHp = null, uint? actorMaxHp = null)
+        {
+            string? text = Config.GetText();
+            if (!Config.Enabled || text == null)
+            {
+                return;
+            }
+
+            DrawLabel(text, origin, parentSize ?? Vector2.Zero, null);
+        }
+
+        protected override void DrawLabel(string text, Vector2 parentPos, Vector2 parentSize, GameObject? actor = null)
+        {
+            ImGui.PushFont(UiBuilder.IconFont);
+            Vector2 size = ImGui.CalcTextSize(text);
+            Vector2 pos = Utils.GetAnchoredPosition(Utils.GetAnchoredPosition(parentPos + Config.Position, -parentSize, Config.FrameAnchor), size, Config.TextAnchor);
+            ImGui.PopFont();
+
+            DrawHelper.DrawInWindow(ID, pos, size, false, true, (drawList) =>
+            {
+                ImGui.PushFont(UiBuilder.IconFont);
+                PluginConfigColor? color = Color(actor);
+
+                if (Config.ShowShadow)
+                {
+                    DrawHelper.DrawShadowText(text, pos, color.Base, Config.ShadowColor.Base, drawList, Config.ShadowOffset);
+                }
+
+                if (Config.ShowOutline)
+                {
+                    DrawHelper.DrawOutlinedText(text, pos, color.Base, Config.OutlineColor.Base, drawList);
+                }
+
+                if (!Config.ShowOutline && !Config.ShowShadow)
+                {
+                    drawList.AddText(pos, color.Base, text);
+                }
+
+                ImGui.PopFont();
+            });
         }
     }
 }